move 'unary' tests to separate file
