+++ types/ramda/tsconfig.json
@@ -227,6 +227,7 @@
         "test/traverse-tests.ts",
         "test/tryCatch-tests.ts",
         "test/type-tests.ts",
-        "test/unapply-tests.ts"
+        "test/unapply-tests.ts",
+        "test/unary-tests.ts"
     ]
 }
